完善文件打开入口与剖面分类交互，升级点云精度链路
- 统一 Open 入口支持点云/航线/工程文件并重构左侧视图工具条\n- 剖面分类新增矩形/多边形模式与状态文案联动，修复工具条提示与分类模式下拉框黑底\n- 点云数据与渲染/业务链路坐标精度适配，补齐中文翻译并通过构建验证

+++ src/domain/AnalysisData.cpp
@@ -29,9 +29,9 @@ QJsonObject pointRecordToJson(const PointRecord& point)
 PointRecord pointRecordFromJson(const QJsonObject& object)
 {
     PointRecord point;
-    point.x = static_cast<float>(object.value(QStringLiteral("x")).toDouble());
-    point.y = static_cast<float>(object.value(QStringLiteral("y")).toDouble());
-    point.z = static_cast<float>(object.value(QStringLiteral("z")).toDouble());
+    point.x = object.value(QStringLiteral("x")).toDouble();
+    point.y = object.value(QStringLiteral("y")).toDouble();
+    point.z = object.value(QStringLiteral("z")).toDouble();
     point.r = static_cast<std::uint8_t>(object.value(QStringLiteral("r")).toInt(255));
     point.g = static_cast<std::uint8_t>(object.value(QStringLiteral("g")).toInt(255));
     point.b = static_cast<std::uint8_t>(object.value(QStringLiteral("b")).toInt(255));
